Refactor: Move sequencer to its own module
