+++ src/player/audio/midisource.rs
@@ -1,6 +1,8 @@
-use midi_msg::{Division, Meta, MidiFile, MidiMsg, TimeCodeType, TrackEvent};
+use midi_msg::MidiFile;
 use rustysynth::{SoundFont, Synthesizer, SynthesizerSettings};
 use std::{sync::Arc, time::Duration};
+
+use super::sequencer::Sequencer;
 
 const SAMPLERATE: u32 = 44100;
 
@@ -101,267 +103,5 @@ impl rodio::Source for MidiSource {
         //self.sequencer
         //    .get_midi_file()
         //    .map(|midifile| Duration::from_secs_f64(midifile.get_length()))
-    }
-}
-
-/// MIDI Sequencer
-struct Sequencer {
-    synthesizer: Synthesizer,
-    midifile: Option<MidiFile>,
-    bpm: f64,
-    /// Index of next event for each track
-    track_positions: Vec<usize>,
-    /// Song position
-    tick: usize,
-    tick_timer: Duration,
-    song_length: Duration,
-}
-impl Sequencer {
-    pub const fn new(synthesizer: Synthesizer) -> Self {
-        Self {
-            synthesizer,
-            midifile: None,
-            bpm: 120.,
-            track_positions: vec![],
-            tick: 0,
-            tick_timer: Duration::ZERO,
-            song_length: Duration::ZERO,
-        }
-    }
-
-    /// Are there no more messages left?
-    pub fn end_of_sequence(&self) -> bool {
-        let Some(midifile) = &self.midifile else {
-            println!("bailed: no midi");
-            return true;
-        };
-        for (i, track) in midifile.tracks.iter().enumerate() {
-            if self.track_positions[i] < track.events().len() {
-                return false;
-            }
-        }
-        println!("bailed: end reached");
-        for (i, track) in midifile.tracks.iter().enumerate() {
-            println!(
-                "Track {i:02?} - len: {} pos: {}",
-                track.events().len(),
-                self.track_positions[i]
-            );
-        }
-        true
-    }
-
-    pub fn play(&mut self, midifile: MidiFile) {
-        self.tick = 0;
-        self.track_positions = vec![0; midifile.tracks.len()];
-        self.midifile = Some(midifile);
-
-        self.synthesizer.reset();
-        self.update_song_length();
-    }
-
-    pub fn render(&mut self) -> [f32; 2] {
-        let Some(events) = self.get_events() else {
-            return [0., 0.];
-        };
-
-        self.tick_timer +=
-            Duration::from_secs_f64(1. / f64::from(self.synthesizer.get_sample_rate()));
-        let tick_duration = self.get_tick_duration();
-        if self.tick_timer >= tick_duration {
-            self.tick_timer -= tick_duration;
-            self.tick += 1;
-        }
-
-        for event in events {
-            match event.event {
-                MidiMsg::ChannelVoice { .. }
-                | MidiMsg::RunningChannelVoice { .. }
-                | MidiMsg::ChannelMode { .. }
-                | MidiMsg::RunningChannelMode { .. } => self.handle_channel_event(&event),
-                midi_msg::MidiMsg::Meta { msg } => self.handle_meta_event(&msg),
-                _ => (),
-            }
-        }
-
-        let mut left = [0.];
-        let mut right = [0.];
-        self.synthesizer.render(&mut left, &mut right);
-        [left[0], right[0]]
-    }
-
-    fn get_events(&mut self) -> Option<Vec<TrackEvent>> {
-        let Some(midifile) = &self.midifile else {
-            return None;
-        };
-
-        let mut events = vec![];
-        for (i, track) in midifile.tracks.iter().enumerate() {
-            loop {
-                let event_idx = self.track_positions[i];
-                if event_idx >= track.len() {
-                    break;
-                }
-                let event = &track.events()[event_idx];
-                let event_tick = midifile
-                    .header
-                    .division
-                    .beat_or_frame_to_tick(event.beat_or_frame)
-                    as usize;
-                if self.tick >= event_tick {
-                    self.track_positions[i] += 1;
-                    events.push(event.clone());
-                    if self.tick > event_tick {
-                        let late = self.tick - event_tick;
-                        println!("Somehow an event was missed! Playing it late ({late} ticks). {event:?}");
-                    }
-                } else {
-                    break;
-                }
-            }
-        }
-        Some(events)
-    }
-
-    fn send_raw_event(&mut self, raw: &[u8]) {
-        let channel = raw[0] & 0x0f;
-        let command = raw[0] & 0xf0;
-        let data1;
-        let data2;
-        match raw.len() {
-            2 => {
-                data1 = raw[1];
-                data2 = 0;
-            }
-            3 => {
-                data1 = raw[1];
-                data2 = raw[2];
-            }
-            _ => panic!("This shouldn't happen. Check length before calling."),
-        }
-        self.synthesizer.process_midi_message(
-            channel.into(),
-            command.into(),
-            data1.into(),
-            data2.into(),
-        );
-    }
-
-    fn handle_channel_event(&mut self, event: &TrackEvent) {
-        let raw = event.event.to_midi();
-
-        if let 2..=3 = raw.len() {
-            self.send_raw_event(&raw);
-            return;
-        }
-
-        if raw.len() == 5 {
-            // Break a message that contains MSB and LSB into two separate ones
-            if raw[1] == 0x64 {
-                let lsb = &raw[0..3];
-                let msb = vec![raw[0], raw[3], raw[4]];
-                self.send_raw_event(lsb);
-                self.send_raw_event(&msb);
-                return;
-            }
-        }
-
-        println!("Unhandled event: raw: {raw:02X?}, event: {event:?}");
-    }
-
-    fn handle_meta_event(&mut self, msg: &Meta) {
-        if let Meta::SetTempo(tempo) = msg {
-            self.bpm = 60_000_000. / f64::from(*tempo);
-        }
-    }
-
-    fn get_tick_duration(&self) -> Duration {
-        let Some(midifile) = &self.midifile else {
-            return Duration::ZERO;
-        };
-        let in_secs = match midifile.header.division {
-            Division::TicksPerQuarterNote(ticks) => 60. / self.bpm / f64::from(ticks),
-            Division::TimeCode {
-                frames_per_second,
-                ticks_per_frame,
-            } => {
-                let fps = match frames_per_second {
-                    TimeCodeType::FPS24 => 24.,
-                    TimeCodeType::FPS25 => 25.,
-                    TimeCodeType::DF30 | TimeCodeType::NDF30 => 30.,
-                };
-                1. / fps / f64::from(ticks_per_frame)
-            }
-        };
-        Duration::from_secs_f64(in_secs)
-    }
-
-    fn update_song_length(&mut self) {
-        let Some(midifile) = &self.midifile else {
-            self.song_length = Duration::ZERO;
-            return;
-        };
-
-        let mut track_positions = vec![0; midifile.tracks.len()];
-        let mut tick = 0;
-        let mut duration = Duration::ZERO;
-        let mut bpm = 120.;
-        loop {
-            let mut done = true;
-            for (i, track) in midifile.tracks.iter().enumerate() {
-                loop {
-                    let event_idx = track_positions[i];
-                    if event_idx >= track.len() {
-                        break;
-                    }
-                    done = false;
-
-                    let event = &track.events()[event_idx];
-                    let event_tick = midifile
-                        .header
-                        .division
-                        .beat_or_frame_to_tick(event.beat_or_frame)
-                        as usize;
-                    if tick >= event_tick {
-                        track_positions[i] += 1;
-                        match &event.event {
-                            MidiMsg::Meta { msg } => match msg {
-                                Meta::SetTempo(tempo) => {
-                                    bpm = 60_000_000. / f64::from(*tempo);
-                                }
-                                _ => (),
-                            },
-                            _ => (),
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            }
-            tick += 1;
-            let tick_duration = match midifile.header.division {
-                Division::TicksPerQuarterNote(ticks) => 60. / bpm / f64::from(ticks),
-                Division::TimeCode {
-                    frames_per_second,
-                    ticks_per_frame,
-                } => {
-                    let fps = match frames_per_second {
-                        TimeCodeType::FPS24 => 24.,
-                        TimeCodeType::FPS25 => 25.,
-                        TimeCodeType::DF30 | TimeCodeType::NDF30 => 30.,
-                    };
-                    1. / fps / f64::from(ticks_per_frame)
-                }
-            };
-            duration += Duration::from_secs_f64(tick_duration);
-            if done {
-                break;
-            }
-        }
-        self.song_length = duration;
-    }
-
-    pub const fn get_song_length(&self) -> Duration {
-        self.song_length
     }
 }
